git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/llvm-utils@40325 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/smt/FormulaManagerView.java
@@ -498,7 +498,7 @@ public class FormulaManagerView {
     } else if (formulaType.isFloatingPointType()) {
       t =
           getFloatingPointFormulaManager()
-              .makeNumber(value, (FormulaType.FloatingPointType) formulaType);
+              .makeNumber((double) value, (FormulaType.FloatingPointType) formulaType);
     } else {
       throw new IllegalArgumentException("Not supported interface");
     }
